pyproject.toml
[project]
name = "poly-maker"
version = "0.1.0"
description = "A market making bot for Polymarket prediction markets"
readme = "README.md"
requires-python = ">=3.9.10"
license = { text = "MIT" }

dependencies = [
    "py-clob-client==0.28.0",
    "python-dotenv==1.2.1",
    "pandas==2.3.3",
    "gspread==6.2.1",
    "gspread-dataframe==4.0.0",
    "sortedcontainers==2.4.0",
    "eth-account==0.13.7",
    "eth-utils==5.3.1",
    "poly_eip712_structs==0.0.1",
    "py_order_utils==0.3.2",
    "requests==2.32.5",
    "websockets==15.0.1",
    "cryptography==46.0.3",
    "google-auth==2.42.1",
    "web3==7.14.0",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.35",
    "langchain-core>=0.3.79",
    "tabulate>=0.9.0",
]

[project.optional-dependencies]
dev = [
    "black==24.4.2",
    "pytest==8.2.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["poly_data", "poly_stats", "poly_utils", "data_updater"]

[tool.black]
line-length = 100
target-version = ["py39"]
